+++ doc/source/release/release_notes.rst
@@ -7,12 +7,28 @@
 Release Notes
 *************
 
+.. release:: 1.1.0
+    :date: 2021-09-08
 
-.. release:: Upcoming
+    .. change:: Changed
 
-    .. change:: Fix
+        Update hook for application launcher.
+
+    .. change:: Fixed
 
         Animation import settings are set by file format.
+
+    .. change:: changed
+        :tags: Setup
+
+        Provide dependency to ftrack-connector-legacy module.
+
+
+.. warning::
+
+    From this version the support for ftrack-connect 1.X is dropped, and
+    only ftrack-conenct 2.0 will be supported up to the integration EOL.
+
 
 .. release:: 1.0.1
     :date: 2020-09-10